requirements.txt
# Requirements file generated from pyproject.toml
# Core data science
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.4.0,<1.5.0
joblib>=1.2.0,<1.4.0

# AutoML / low-code
pycaret>=3.3.0,<4.0.0

# Experiment tracking & orchestration
mlflow>=2.8.0,<3.0.0
hydra-core>=1.3.0,<2.0.0
dvc>=3.0.0,<4.0.0

# App / dashboard
streamlit>=1.28.0,<2.0.0

# Monitoring
evidently>=0.4.0,<1.0.0

# Notebooks & visualization
jupyter>=1.0.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.17.0,<6.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0
tqdm>=4.66.0

# Dev dependencies
pytest>=7.4.0
flake8>=6.1.0
mypy>=1.5.0
black>=23.9.0
